assignment 6

+++ cst438_registration/src/main/java/com/cst438/controller/ScheduleController.java
@@ -21,7 +21,12 @@ import com.cst438.domain.EnrollmentRepository;
 import com.cst438.domain.ScheduleDTO;
 import com.cst438.domain.Student;
 import com.cst438.domain.StudentRepository;
+import com.cst438.domain.User;
+import com.cst438.domain.UserRepository;
 import com.cst438.service.GradebookService;
+import java.security.Principal;
+
+
 @RestController
 @CrossOrigin 
 public class ScheduleController {
@@ -37,22 +42,31 @@ public class ScheduleController {
 	
 	@Autowired
 	GradebookService gradebookService;
+	
+	@Autowired
+	UserRepository userRespository;
+	
 	/*
 	 * get current schedule for student.
 	 */
 	@GetMapping("/schedule")
-	public ScheduleDTO[] getSchedule( @RequestParam("year") int year, @RequestParam("semester") String semester ) {
-		System.out.println("/schedule called.");
-		String student_email = "test@csumb.edu";   // student's email 
-		
-		Student student = studentRepository.findByEmail(student_email);
-		if (student != null) {
-			System.out.println("/schedule student "+student.getName()+" "+student.getStudent_id());
-			List<Enrollment> enrollments = enrollmentRepository.findStudentSchedule(student_email, year, semester);
-			ScheduleDTO[] sched = createSchedule(year, semester, student, enrollments);
-			return sched;
+	public ScheduleDTO[] getSchedule(Principal p, @RequestParam("year") int year, @RequestParam("semester") String semester ) {
+		User user = userRespository.findByUsername(p.getName());
+		if(user.getRole().equals("USER")) {
+			System.out.println("/schedule called.");
+			String student_email = "test@csumb.edu";   // student's email 
+			
+			Student student = studentRepository.findByEmail(student_email);
+			if (student != null) {
+				System.out.println("/schedule student "+student.getName()+" "+student.getStudent_id());
+				List<Enrollment> enrollments = enrollmentRepository.findStudentSchedule(student_email, year, semester);
+				ScheduleDTO[] sched = createSchedule(year, semester, student, enrollments);
+				return sched;
+			} else {
+				return new ScheduleDTO[0];   // return empty schedule for unknown student.
+			}
 		} else {
-			return new ScheduleDTO[0];   // return empty schedule for unknown student.
+			return null;
 		}
 	}
 	/*
@@ -60,44 +74,52 @@ public class ScheduleController {
 	 */
 	@PostMapping("/schedule/course/{id}")
 	@Transactional
-	public ScheduleDTO addCourse( @PathVariable int id  ) { 
-		String student_email = "test@csumb.edu";   // student's email 
-		Student student = studentRepository.findByEmail(student_email);
-		Course course  = courseRepository.findById(id).orElse(null);
-		// student.status
-		// = 0  ok to register.  != 0 registration is on hold.		
-		if (student!= null && course!=null && student.getStatusCode()==0) {
-			// TODO check that today's date is not past add deadline for the course.
-			Enrollment enrollment = new Enrollment();
-			enrollment.setStudent(student);
-			enrollment.setCourse(course);
-			enrollment.setYear(course.getYear());
-			enrollment.setSemester(course.getSemester());
-			enrollmentRepository.save(enrollment);
-			// notify grade book of new enrollment event
-			gradebookService.enrollStudent(student_email, student.getName(), course.getCourse_id());
-			ScheduleDTO result = createSchedule(enrollment);
-			return result;
+	public ScheduleDTO addCourse(Principal p, @PathVariable int id  ) { 
+		User user = userRespository.findByUsername(p.getName());
+		if(user.getRole().equals("USER")) {
+			String student_email = "test@csumb.edu";   // student's email 
+			Student student = studentRepository.findByEmail(student_email);
+			Course course  = courseRepository.findById(id).orElse(null);
+			// student.status
+			// = 0  ok to register.  != 0 registration is on hold.		
+			if (student!= null && course!=null && student.getStatusCode()==0) {
+				// TODO check that today's date is not past add deadline for the course.
+				Enrollment enrollment = new Enrollment();
+				enrollment.setStudent(student);
+				enrollment.setCourse(course);
+				enrollment.setYear(course.getYear());
+				enrollment.setSemester(course.getSemester());
+				enrollmentRepository.save(enrollment);
+				// notify grade book of new enrollment event
+				gradebookService.enrollStudent(student_email, student.getName(), course.getCourse_id());
+				ScheduleDTO result = createSchedule(enrollment);
+				return result;
+			} else {
+				throw  new ResponseStatusException( HttpStatus.BAD_REQUEST, "Course_id invalid or student not allowed to register for the course.  "+id);
+			}
 		} else {
-			throw  new ResponseStatusException( HttpStatus.BAD_REQUEST, "Course_id invalid or student not allowed to register for the course.  "+id);
-		}	
+			throw  new ResponseStatusException( HttpStatus.BAD_REQUEST, "You dont have access  ");
+		}
 	}
 	/*
 	 * drop a course from student schedule
 	 */
 	@DeleteMapping("/schedule/{enrollment_id}")
 	@Transactional
-	public void dropCourse(  @PathVariable int enrollment_id  ) {
-		String student_email = "test@csumb.edu";   // student's email 
-		// TODO  check that today's date is not past deadline to drop course.
-		Enrollment enrollment = enrollmentRepository.findById(enrollment_id).orElse(null);
-		// verify that student is enrolled in the course.
-		if (enrollment!=null && enrollment.getStudent().getEmail().equals(student_email)) {
-			// OK.  drop the course.
-			 enrollmentRepository.delete(enrollment);
-		} else {
-			// something is not right with the enrollment.  
-			throw  new ResponseStatusException( HttpStatus.BAD_REQUEST, "Enrollment_id invalid. "+enrollment_id);
+	public void dropCourse(Principal p, @PathVariable int enrollment_id  ) {
+		User user = userRespository.findByUsername(p.getName());
+		if(user.getRole().equals("USER")) {		
+			String student_email = "test@csumb.edu";   // student's email 
+			// TODO  check that today's date is not past deadline to drop course.
+			Enrollment enrollment = enrollmentRepository.findById(enrollment_id).orElse(null);
+			// verify that student is enrolled in the course.
+			if (enrollment!=null && enrollment.getStudent().getEmail().equals(student_email)) {
+				// OK.  drop the course.
+				 enrollmentRepository.delete(enrollment);
+			} else {
+				// something is not right with the enrollment.  
+				throw  new ResponseStatusException( HttpStatus.BAD_REQUEST, "Enrollment_id invalid. "+enrollment_id);
+			}
 		}
 	}
